requirements.txt
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
matplotlib==3.7.1